Use configd_sess_setup_shared() to create session
The existing configd_sess_setup() API now creates a session which is
private to the requesting user. NETCONF semantics require a session
which is shared among users, therefore update to use the new API for
shared session creation.

We no longer explicitly check if the session exists before setting it
up as configd_sess_setup_shared() handles this gracefully. Also in
the (unlikely) case where "NETCONF" already exists as a private
session this ensures we log a descriptive error message.

VRVDR-53078