Add support for CopyConfig
Part of VRVDR-46458.

Add support for NETCONF copy-config operation with source specified as XML (inside <config> tags) and target as 'candidate' datastore.  Needed to pass incoming rpc on to the existing copy_config handler function, and modify the latter to support the new variant of the copy-config operation.

UT updated to verify error handling.