Reject copy-config requests with same source and target datastores
Fixes: VRVDR-54090

The NETCONF RFC 6241 mandates that an invalid-value error be generated if the source and target fields in a copy-config operation are the same.  While this is checked in configd, netconfd treats such a request as a no-op, returning ok rather than returning an error or passing the request on to configd.  Fix is to now reject such requests with an invalid-value error tag.  This is done for candidate, startup and running datastores, verified by acceptance test (separate commit).

(cherry picked from commit e49aae42f129544dd76cf94af4aa41ebd8a28092)